# Trucking-Game-API/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Trucking Software Public API",
		"description": "API for trucking software web application",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3001/api/v1",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/tickets/incomplete": {
			"get": {
				"description": "Get paginated list of incomplete tickets",
				"summary": "Get paginated list of incompleted tickets",
				"operationId": "getIncompleteTickets",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					}
				],
				"responses": {
					"200": {
						"description": "Get incomplete tickets success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TicketPaginatedResponse"
								}
							}
						}
					},
					"400": {
						"description": "Get incomplete tickets fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/completed": {
			"get": {
				"description": "Get paginated list of completed tickets",
				"summary": "Get paginated list of completed tickets",
				"operationId": "getCompletedTickets",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					}
				],
				"responses": {
					"200": {
						"description": "Get completed tickets success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TicketPaginatedResponse"
								}
							}
						}
					},
					"400": {
						"description": "Get completed tickets fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/unassigned": {
			"get": {
				"description": "Get a paginated list of unassigned tickets",
				"summary": "Get a paginated list of unassigned tickets",
				"operationId": "getUnassignedTickets",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					}
				],
				"responses": {
					"200": {
						"description": "Get unassigned tickets success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TicketPaginatedResponse"
								}
							}
						}
					},
					"400": {
						"description": "Get unassigned tickets fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/mine": {
			"get": {
				"description": "Get a list of tickets belonging to a logged in user",
				"summary": "Get a list of tickets belonging to a logged in user",
				"operationId": "getTicketsByUser",
				"tags": ["tickets"],
				"responses": {
					"200": {
						"description": "Get my tickets success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"tickets": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Ticket"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get my tickets fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/{ticketId}": {
			"parameters": [
				{
					"in": "path",
					"name": "ticketId",
					"schema": {
						"type": "string"
					},
					"required": true,
					"description": "Id of a specific Ticket document"
				}
			],
			"get": {
				"description": "Get a ticket by id",
				"summary": "Get a ticket by id",
				"operationId": "getTicketById",
				"tags": ["tickets"],
				"responses": {
					"200": {
						"description": "Get ticket by id success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"ticket": {
											"$ref": "#/components/schemas/Ticket"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get ticket by id fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ticket"
								}
							}
						}
					}
				}
			},
			"delete": {
				"description": "Delete a single ticket by id",
				"summary": "Delete a single ticket by id",
				"operationId": "deleteOneTicket",
				"tags": ["tickets"],
				"responses": {
					"200": {
						"description": "Delete ticket success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"messsage": {
											"type": "string",
											"example": "Successfully deleted ticket"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete ticket fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/byUser/{email}": {
			"get": {
				"description": "Get all tickets created by a specific email",
				"summary": "Get all tickets created by a specific email",
				"operationId": "getTicketsByCreator",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"schema": {
							"type": "string",
							"example": "test@email.com"
						},
						"required": true,
						"description": "Email for the creator of tickets to search for"
					}
				],
				"responses": {
					"200": {
						"description": "Get tickets by creator success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"tickets": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Ticket"
											}
										},
										"totalDocuments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get tickets by creator fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/byAssignee/{adminEmail}": {
			"get": {
				"description": "Get all tickets assigned to a specific email",
				"summary": "Get all tickets assigned to a specific email",
				"operationId": "getTicketsByAssignee",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "path",
						"name": "adminEmail",
						"schema": {
							"type": "string",
							"example": "morgan@email.com"
						},
						"required": true,
						"description": "Email for the assignee of tickets to search for"
					}
				],
				"responses": {
					"200": {
						"description": "Get tickets by assignee success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"tickets": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Ticket"
											}
										},
										"totalDocuments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get tickets by assignee fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/{ticketId}/comments": {
			"get": {
				"description": "Get comments for a specific ticket",
				"summary": "Get comments for a specific ticket",
				"operationId": "getCommentsByTicket",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "path",
						"name": "ticketId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of a specific ticket"
					}
				],
				"responses": {
					"200": {
						"description": "Get comments by ticket success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"comments": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/TicketComment"
											}
										},
										"totalComments": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get comments by ticket fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets": {
			"post": {
				"description": "Create a new ticket",
				"summary": "Create a new ticket",
				"operationId": "createTicket",
				"tags": ["tickets"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TicketBody"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Create ticket success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"ticket": {
											"$ref": "#/components/schemas/Ticket"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Create ticket fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/{ticketId}/message": {
			"post": {
				"description": "Create a new ticket message",
				"summary": "Create a new ticket message",
				"operationId": "createTicketMessage",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "path",
						"name": "ticketId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of a specific Ticket"
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"text": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Create ticket comment success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"comment": {
											"$ref": "#/components/schemas/TicketComment"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Create ticket comment fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/assignUser": {
			"patch": {
				"description": "Assign a ticket to an admin",
				"summary": "Assign a ticket to an admin",
				"operationId": "assignTicketToAdmin",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "query",
						"name": "ticketId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id of a specific Ticket"
					},
					{
						"in": "query",
						"name": "adminEmail",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Email of a specific admin"
					}
				],
				"responses": {
					"200": {
						"description": "Assign ticket success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"updatedTicket": {
											"$ref": "#/components/schemas/Ticket"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Assign ticket fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/updateStatus/{ticketId}": {
			"patch": {
				"description": "Update status of a single ticket",
				"summary": "Update status of a single ticket",
				"operationId": "updateTicketStatus",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "path",
						"name": "ticketId",
						"schema": {
							"type": "string"
						},
						"required": true,
						"description": "Id for a specific ticket"
					}
				],
				"responses": {
					"200": {
						"description": "Update ticket status success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"updatedTicket": {
											"$ref": "#/components/schemas/Ticket"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Update ticket status fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/tickets/multiple": {
			"delete": {
				"description": "Delete multiple tickets",
				"summary": "Delete multiple tickets",
				"operationId": "deleteMultipleTickets",
				"tags": ["tickets"],
				"parameters": [
					{
						"in": "query",
						"name": "ticketId",
						"schema": {
							"type": "array",
							"items": "string'"
						},
						"required": true,
						"description": "One or more ids of specific tickets"
					}
				],
				"responses": {
					"200": {
						"description": "Delete multiple tickets success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"message": {
											"type": "string",
											"example": "Successfully deleted tickets"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Delete multiple tickets fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteMultipleErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
...
			"DeleteMultipleErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"message": {
						"type": "string",
						"example": "One or more artcles was not found"
					},
					"incomplete": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"error": {
						"type": "string"
					}
				}
			},
...
			"Ticket": {
				"type": "object",
				"properties": {
					"ticketNumber": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "integer"
					},
					"category": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"assignedTo": {
						"$ref": "#/components/schemas/User"
					},
					"comments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TicketComment"
						}
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				}
			},
			"TicketBody": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"category": {
						"type": "string"
					}
				},
				"required": ["message", "email", "category"]
			},
			"TicketComment": {
				"type": "object",
				"properties": {
					"text": {
						"type": "string"
					},
					"relatedTicket": {
						"$ref": "#/components/schemas/Ticket"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
			"TicketPaginatedResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": true
					},
					"tickets": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Ticket"
						}
					},
					"totalDocuments": {
						"type": "integer"
					}
				}
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "integer"
					},
					"phone": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					},
					"hasPaid": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	]
}